+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ```
     * Removed the version paramter from the main file. A getter method was created, who get the version from config.xml.
     * For the QA site add a specific parameter when call Simply Connect.
+    * Fixed the check if the plugin is configured.
 ```
 
 # 2.1.0